Reorder laser cutter training
Hazards should come first, especially because the training starts with the
safety section so it's just a repeat. Make these messages the same to avoid
repeating things in different ways.

Move stopping the laser to later, so that the training can be completed in
order.

Make the basic maintenance mandatory and move alignment to the intermediate
section which is now marked as optional.